requirements.txt
openai>=1.30.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
tiktoken>=0.7.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pypdf>=3.17.0
python-docx>=1.1.0
